正式比赛实用程序
最终正式比赛的采用的的程序，能够正确的识别走线，但是速度很低，方向控制依然存在问题，反光情况没能考虑完善——我们尽力了

+++ Board/inc/XK_motion.h
@@ -11,6 +11,6 @@
 #ifndef _XK_MOTION_H_
 #define _XK_MOTION_H_
 
-extern void motion(uint8 left, uint8 right);
+extern void motion(float left, float right);
 
 #endif  //__XK_MOTION_H__
